Remove end-to-end tests
Other than the workflow, which is removed, they're now duplicated in the unit tests
